# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SkillRewind API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rebuilds/{rebuild_id}": {
      "get": {
        "operationId": "get_rebuild_api_v1_rebuilds__rebuild_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rebuild_id",
            "required": true,
            "schema": {
              "title": "Rebuild Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Rebuild Api V1 Rebuilds  Rebuild Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rebuild",
        "tags": [
          "rebuilds"
        ]
      }
    },
    "/api/v1/rebuilds/{rebuild_id}/exclusions": {
      "get": {
        "operationId": "get_rebuild_exclusions_api_v1_rebuilds__rebuild_id__exclusions_get",
        "parameters": [
          {
            "in": "path",
            "name": "rebuild_id",
            "required": true,
            "schema": {
              "title": "Rebuild Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Rebuild Exclusions Api V1 Rebuilds  Rebuild Id  Exclusions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rebuild Exclusions",
        "tags": [
          "rebuilds"
        ]
      }
    },
    "/api/v1/rebuilds/{rebuild_id}/output": {
      "get": {
        "operationId": "get_rebuild_output_api_v1_rebuilds__rebuild_id__output_get",
        "parameters": [
          {
            "in": "path",
            "name": "rebuild_id",
            "required": true,
            "schema": {
              "title": "Rebuild Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Rebuild Output Api V1 Rebuilds  Rebuild Id  Output Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rebuild Output",
        "tags": [
          "rebuilds"
        ]
      }
    },
    "/api/v1/rebuilds/{rebuild_id}/support": {
      "get": {
        "operationId": "get_rebuild_support_api_v1_rebuilds__rebuild_id__support_get",
        "parameters": [
          {
            "in": "path",
            "name": "rebuild_id",
            "required": true,
            "schema": {
              "title": "Rebuild Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Rebuild Support Api V1 Rebuilds  Rebuild Id  Support Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rebuild Support",
        "tags": [
          "rebuilds"
        ]
      }
    },
    "/api/v1/rebuilds/{rebuild_id}/verification": {
      "get": {
        "operationId": "get_rebuild_verification_api_v1_rebuilds__rebuild_id__verification_get",
        "parameters": [
          {
            "in": "path",
            "name": "rebuild_id",
            "required": true,
            "schema": {
              "title": "Rebuild Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Rebuild Verification Api V1 Rebuilds  Rebuild Id  Verification Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rebuild Verification",
        "tags": [
          "rebuilds"
        ]
      }
    },
...
    "/api/v1/revocations/{revocation_id}/rebuilds": {
      "get": {
        "operationId": "list_rebuilds_for_revocation_api_v1_revocations__revocation_id__rebuilds_get",
        "parameters": [
          {
            "in": "path",
            "name": "revocation_id",
            "required": true,
            "schema": {
              "title": "Revocation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Rebuilds For Revocation Api V1 Revocations  Revocation Id  Rebuilds Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rebuilds For Revocation",
        "tags": [
          "rebuilds"
        ]
      }
    },
...
  }
}